model_tree::attribute split mod
